feat: add confirmation dialog for unsaved changes in SimpleProductTable
