+++ src/components/product/SimpleProductTable.jsx
@@ -11,12 +11,7 @@ import {
     Button,
     Checkbox,
     Typography,
-    IconButton,
-    Dialog,
-    DialogTitle,
-    DialogContent,
-    DialogContentText,
-    DialogActions
+    IconButton
 } from "@mui/material";
 import {
     Edit as EditIcon,
@@ -25,171 +20,43 @@ import {
     ArrowBack as ArrowBackIcon,
     Search as SearchIcon
 } from "@mui/icons-material";
-import {useState, useEffect, useMemo} from "react";
-import {useProduct} from "../../hooks/product/useProduct.js";
-import {toast} from "react-toastify";
+import {useEffect} from "react";
+import {useProductTableSimple} from "../../hooks/product/useProductTableSimple.js";
+import {ConfirmDialog} from "./ConfirmDialog.jsx";
+
 
 export const SimpleProductTable = () => {
-    const {products, handleGetProducts, handleUpdatePriceProduct, handleCancel} = useProduct();
-    const [editingProducts, setEditingProducts] = useState(new Set());
-    const [selectedProducts, setSelectedProducts] = useState(new Set());
-    const [priceChanges, setPriceChanges] = useState({});
-    const [isLoading, setIsLoading] = useState(false);
-    const [nameFilter, setNameFilter] = useState('');
-    const [showConfirmDialog, setShowConfirmDialog] = useState(false);
-
-    // Filter products by name
-    const filteredProducts = useMemo(() => {
-        if (!nameFilter.trim()) {
-            return products;
-        }
-        return products.filter(product =>
-            product.name.toLowerCase().includes(nameFilter.toLowerCase())
-        );
-    }, [products, nameFilter]);
+    const {
+        handleGetProducts,
+        handleBackClick,
+        nameFilter,
+        setNameFilter,
+        selectedProducts,
+        hasUnsavedChanges,
+        modifiedProductsCount,
+        shouldShowSaveButton,
+        handleAutoSaveChanges,
+        isLoading,
+        filteredProducts,
+        handleSelectAll,
+        handleSelectProduct,
+        editingProducts,
+        priceChanges,
+        handlePriceChange,
+        handleSaveSinglePrice,
+        handleCancelEdit,
+        handleEditPrice,
+        showConfirmDialog,
+        handleCancelBack,
+        handleConfirmBack
+    } = useProductTableSimple();
 
     useEffect(() => {
         handleGetProducts();
     }, []);
 
-    const handleSelectProduct = (productId) => {
-        const newSelected = new Set(selectedProducts);
-        if (newSelected.has(productId)) {
-            newSelected.delete(productId);
-        } else {
-            newSelected.add(productId);
-        }
-        setSelectedProducts(newSelected);
-    };
-
-    const handleSelectAll = () => {
-        if (selectedProducts.size === filteredProducts.length) {
-            setSelectedProducts(new Set());
-        } else {
-            setSelectedProducts(new Set(filteredProducts.map(p => p.id)));
-        }
-    };
-
-    const handleEditPrice = (productId) => {
-        const newEditing = new Set(editingProducts);
-        newEditing.add(productId);
-        setEditingProducts(newEditing);
-
-        // Initialize price change with current price
-        const product = products.find(p => p.id === productId);
-        setPriceChanges(prev => ({
-            ...prev,
-            [productId]: product.price
-        }));
-    };
-
-    const handleCancelEdit = (productId) => {
-        const newEditing = new Set(editingProducts);
-        newEditing.delete(productId);
-        setEditingProducts(newEditing);
-
-        // Remove price change
-        const newPriceChanges = {...priceChanges};
-        delete newPriceChanges[productId];
-        setPriceChanges(newPriceChanges);
-    };
-
-    const handlePriceChange = (productId, newPrice) => {
-        setPriceChanges(prev => ({
-            ...prev,
-            [productId]: parseFloat(newPrice) || 0
-        }));
-    };
-
-    const handleSaveSinglePrice = async (productId) => {
-        const newPrice = priceChanges[productId];
-
-        if (newPrice <= 0) {
-            toast.error('El precio debe ser mayor a 0');
-            return;
-        }
-
-        const success = await handleUpdatePriceProduct(productId, newPrice);
-        if (success) {
-            handleCancelEdit(productId);
-        }
-    };
-
-    // Check if there are unsaved changes
-    const hasUnsavedChanges = useMemo(() => {
-        return Object.keys(priceChanges).some(productId => {
-            const product = products.find(p => p.id === parseInt(productId));
-            return product && priceChanges[productId] !== product.price;
-        });
-    }, [priceChanges, products]);
-
-    // Count modified products
-    const modifiedProductsCount = useMemo(() => {
-        return Object.keys(priceChanges).filter(productId => {
-            const product = products.find(p => p.id === parseInt(productId));
-            return product && priceChanges[productId] !== product.price && priceChanges[productId] > 0;
-        }).length;
-    }, [priceChanges, products]);
-
-    // Auto-enable save button when more than 2 modifications
-    const shouldShowSaveButton = modifiedProductsCount >= 2;
-
-    const handleBackClick = () => {
-        if (hasUnsavedChanges) {
-            setShowConfirmDialog(true);
-        } else {
-            handleCancel();
-        }
-    };
-
-    const handleConfirmBack = () => {
-        setShowConfirmDialog(false);
-        handleCancel();
-    };
-
-    const handleCancelBack = () => {
-        setShowConfirmDialog(false);
-    };
-
-    const handleAutoSaveChanges = async () => {
-        const modifiedProducts = Object.keys(priceChanges).filter(productId => {
-            const product = products.find(p => p.id === parseInt(productId));
-            return product && priceChanges[productId] !== product.price && priceChanges[productId] > 0;
-        });
-
-        if (modifiedProducts.length === 0) {
-            toast.error('No hay cambios válidos para guardar');
-            return;
-        }
-
-        setIsLoading(true);
-        let successCount = 0;
-
-        for (const productId of modifiedProducts) {
-            const newPrice = priceChanges[productId];
-            const success = await handleUpdatePriceProduct(parseInt(productId), newPrice);
-            if (success) {
-                successCount++;
-            }
-        }
-
-        setIsLoading(false);
-
-        if (successCount > 0) {
-            toast.success(`${successCount} precios actualizados exitosamente`);
-            // Clear the changes after successful save
-            const clearedChanges = {...priceChanges};
-            modifiedProducts.forEach(productId => {
-                delete clearedChanges[productId];
-            });
-            setPriceChanges(clearedChanges);
-            setEditingProducts(new Set());
-        }
-    };
-
     return (
         <Box sx={{width: '100%'}}>
-            {/* Back button */}
             <Box sx={{mb: 2}}>
                 <Button
                     variant="outlined"
@@ -205,7 +72,6 @@ export const SimpleProductTable = () => {
                 Modificación de Precios
             </Typography>
 
-            {/* Search bar */}
             <Box sx={{mb: 2}}>
                 <TextField
                     variant="outlined"
@@ -222,7 +88,6 @@ export const SimpleProductTable = () => {
                 />
             </Box>
 
-            {/* Bulk actions */}
             <Box sx={{mb: 2, display: 'flex', gap: 2, flexWrap: 'wrap', alignItems: 'center'}}>
                 <Typography variant="body2">
                     {selectedProducts.size} productos seleccionados
@@ -234,7 +99,6 @@ export const SimpleProductTable = () => {
                     </Typography>
                 )}
 
-                {/* Auto-save button when 2+ modifications */}
                 {shouldShowSaveButton && (
                     <Button
                         variant="contained"
@@ -333,32 +197,8 @@ export const SimpleProductTable = () => {
                     </TableBody>
                 </Table>
             </TableContainer>
-
-            {/* Confirmation Dialog */}
-            <Dialog
-                open={showConfirmDialog}
-                onClose={handleCancelBack}
-                aria-labelledby="confirm-dialog-title"
-                aria-describedby="confirm-dialog-description"
-            >
-                <DialogTitle id="confirm-dialog-title">
-                    Cambios sin guardar
-                </DialogTitle>
-                <DialogContent>
-                    <DialogContentText id="confirm-dialog-description">
-                        Tienes {modifiedProductsCount} cambios sin guardar. ¿Estás seguro de que quieres salir sin
-                        guardar los cambios?
-                    </DialogContentText>
-                </DialogContent>
-                <DialogActions>
-                    <Button onClick={handleCancelBack} color="primary">
-                        Cancelar
-                    </Button>
-                    <Button onClick={handleConfirmBack} color="error" autoFocus>
-                        Salir sin guardar
-                    </Button>
-                </DialogActions>
-            </Dialog>
+            <ConfirmDialog showConfirmDialog={showConfirmDialog} handleCancelBack={handleCancelBack}
+                           handleConfirmBack={handleConfirmBack} modifiedProductsCount={modifiedProductsCount}/>
         </Box>
     );
 }
